pyproject.toml
[tool.poetry]
name = "finalproject-konovalova-m25-555"
version = "0.1.0"
description = "Платформа для отслеживания и симуляции торговли валютами"
authors = ["Elizaveta Konovalova <Eliza050203@yandex.ru>"]
readme = "README.md"

packages = [
    {include = "valutatrade_hub"}
]

[tool.poetry.scripts]
project = "valutatrade_hub.cli.interface:main"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.5"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.6"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["venv", ".venv", "dist", "pycache"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.valutatrade]
data_path = "data"
users_file_path = "users.json"
portfolios_file_path = "portfolios.json"
rates_file_path = "rates.json"
rates_ttl_seconds = "300"
default_base_currency = "USD"
log_level = "INFO"
log_format = "[%(levelname)s] %(asctime)s %(message)s"
date_format = "%Y-%m-%dT%H:%M:%S"
